Added canonicalization of condition number to sets.py
