+++ cvxpy/reductions/dqcp2dcp/sets.py
@@ -15,7 +15,11 @@ limitations under the License.
 """
 from cvxpy import atoms
 from cvxpy.atoms.affine import binary_operators as bin_op
+from cvxpy.atoms.affine.diag import diag_vec
+from cvxpy.atoms.affine.promote import promote
+from cvxpy.constraints.psd import PSD
 from cvxpy.expressions.constants.parameter import Parameter
+from cvxpy.expressions.variable import Variable
 
 
 # Sublevel sets for quasiconvex atoms.
@@ -124,6 +128,22 @@ def gen_lambda_max_sub(expr, t):
             (t * expr.args[1] - expr.args[0] >> 0)]
 
 
+def condition_number_sub(expr, t):
+    A = expr.args[0]
+    n = A.shape[0]
+    u = Variable(pos=True)
+
+    prom_ut = promote(u * t, (n,))
+    prom_u = promote(u, (n,))
+    tmp_expr1 = A - diag_vec(prom_u)
+    tmp_expr2 = diag_vec(prom_ut) - A
+
+    return [A == A.T,
+            A >> 0,
+            PSD(tmp_expr1),
+            PSD(tmp_expr2)]
+
+
 SUBLEVEL_SETS = {
     atoms.multiply: mul_sub,
     bin_op.DivExpression: ratio_sub,
@@ -131,6 +151,7 @@ SUBLEVEL_SETS = {
     atoms.sign: sign_sub,
     atoms.dist_ratio: dist_ratio_sub,
     atoms.gen_lambda_max: gen_lambda_max_sub,
+    atoms.condition_number: condition_number_sub,
 }
